- Merge SimplePersistentVectorStore and InMemoryVector store to be SimpleVectorStore
