+++ spring-ai-core/src/main/java/org/springframework/ai/vectorstore/SimpleVectorStore.java
@@ -1,10 +1,21 @@
 package org.springframework.ai.vectorstore;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.ai.document.Document;
 import org.springframework.ai.embedding.EmbeddingClient;
+import org.springframework.core.io.Resource;
+import org.springframework.util.StreamUtils;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.charset.Charset;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -14,15 +25,15 @@ import java.util.concurrent.ConcurrentHashMap;
  * @author Mark Pollack
  * @author Christian Tzolov
  */
-public class InMemoryVectorStore implements VectorStore {
+public class SimpleVectorStore implements VectorStore {
 
-	private static final Logger logger = LoggerFactory.getLogger(InMemoryVectorStore.class);
+	private static final Logger logger = LoggerFactory.getLogger(SimpleVectorStore.class);
 
 	protected Map<String, Document> store = new ConcurrentHashMap<>();
 
 	protected EmbeddingClient embeddingClient;
 
-	public InMemoryVectorStore(EmbeddingClient embeddingClient) {
+	public SimpleVectorStore(EmbeddingClient embeddingClient) {
 		Objects.requireNonNull(embeddingClient, "EmbeddingClient must not be null");
 		this.embeddingClient = embeddingClient;
 	}
@@ -64,6 +75,81 @@ public class InMemoryVectorStore implements VectorStore {
 			.toList();
 
 		return similarities;
+	}
+
+	/**
+	 * Serialize the vector store content into a file in JSON format.
+	 * @param file the file to save the vector store content
+	 */
+	public void save(File file) {
+		String json = getVectorDbAsJson();
+		try {
+			if (!file.exists()) {
+				logger.info("Creating new vector store file: " + file);
+				file.createNewFile();
+			}
+			else {
+				logger.info("Replacing existing vector store file: " + file);
+				file.delete();
+				file.createNewFile();
+			}
+		}
+		catch (IOException ex) {
+			throw new RuntimeException(ex);
+		}
+		try (OutputStream stream = new FileOutputStream(file)) {
+			StreamUtils.copy(json, Charset.forName("UTF-8"), stream);
+		}
+		catch (IOException e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	/**
+	 * Deserialize the vector store content from a file in JSON format into memory.
+	 * @param file the file to load the vector store content
+	 */
+	public void load(File file) {
+		TypeReference<HashMap<String, Document>> typeRef = new TypeReference<>() {
+		};
+		ObjectMapper objectMapper = new ObjectMapper();
+		try {
+			Map<String, Document> deserializedMap = objectMapper.readValue(file, typeRef);
+			this.store = deserializedMap;
+		}
+		catch (IOException ex) {
+			throw new RuntimeException(ex);
+		}
+	}
+
+	/**
+	 * Deserialize the vector store content from a resource in JSON format into memory.
+	 * @param resource the resource to load the vector store content
+	 */
+	public void load(Resource resource) {
+		TypeReference<HashMap<String, Document>> typeRef = new TypeReference<>() {
+		};
+		ObjectMapper objectMapper = new ObjectMapper();
+		try {
+			Map<String, Document> deserializedMap = objectMapper.readValue(resource.getInputStream(), typeRef);
+			this.store = deserializedMap;
+		}
+		catch (IOException ex) {
+			throw new RuntimeException(ex);
+		}
+	}
+
+	private String getVectorDbAsJson() {
+		ObjectMapper objectMapper = new ObjectMapper();
+		ObjectWriter objectWriter = objectMapper.writerWithDefaultPrettyPrinter();
+		String json;
+		try {
+			json = objectWriter.writeValueAsString(this.store);
+		}
+		catch (JsonProcessingException e) {
+			throw new RuntimeException("Error serializing documentMap to JSON.", e);
+		}
+		return json;
 	}
 
 	private List<Double> getUserQueryEmbedding(String query) {
